requirements.txt
boto3==1.35.0
botocore==1.35.0
celery==5.4.0
Flask==3.0.3
numpy==2.1.1
Pillow==10.4.0
scikit-image==0.23.2
